+++ src/utils/index.ts
@@ -1,2 +1,4 @@
 export * from './with-request-timeout';
 export * from './truncate';
+export * from './get-package-version';
+export * from './check-for-update';